requirements.txt
datasets>=4.0.0
numpy>=2.3.2
python-dotenv>=1.1.1
torch>=2.3.1
wandb
